+++ lib/src/gutter.dart
@@ -10,27 +10,30 @@ class Gutter extends StatelessWidget {
   /// Creates a new [Gutter] widget.
   const Gutter({
     super.key,
-    this.size,
     this.type = GutterType.medium,
+    this.size,
     this.scaleFactor = 1,
   });
 
   /// The spacing used on small screens according to Material Design's
-  static const double materialSpacingSmall = 16.0;
+  static const double materialSpacingSmall = 16;
 
   /// The spacing used on medium and large screens according to Material Design's
-  static const double materialSpacingMediumAndUp = 24.0;
+  static const double materialSpacingMediumAndUp = 24;
+
+  /// The default scale factor for a gutter.
+  static const double scaleFactorDefault = 1;
 
   /// The default scale factor for a small increase.
-  static const int scaleFactorSmallDefault = 2;
+  static const double scaleFactorDefaultSmall = 2;
 
   /// The default scale factor for a medium increase.
-  static const int scaleFactorMediumDefault = 4;
+  static const double scaleFactorDefaultMedium = 4;
 
   /// The type of gutter to create.
   final GutterType type;
 
-  /// The size of the gap.
+  /// The size of the [Gutter].
   final double? size;
 
   /// The scale factor to apply to the gutter size.
@@ -38,59 +41,62 @@ class Gutter extends StatelessWidget {
 
   /// Creates a new [Gutter] widget with a single line size.
   const Gutter.singleLine({super.key})
-      : size = 1,
-        type = GutterType.expand,
+      : type = GutterType.expand,
+        size = 1,
         scaleFactor = 1;
 
   /// Creates a new [Gutter] widget with a tiny size.
   const Gutter.tiny({super.key, this.scaleFactor = 1})
-      : size = null,
-        type = GutterType.tiny;
+      : type = GutterType.tiny,
+        size = null;
 
   /// Creates a new [Gutter] widget with a small size.
   const Gutter.small({super.key, this.scaleFactor = 1})
-      : size = null,
-        type = GutterType.small;
+      : type = GutterType.small,
+        size = null;
 
   /// Creates a new [Gutter] widget with a medium size.
   const Gutter.medium({super.key, this.scaleFactor = 1})
-      : size = null,
-        type = GutterType.medium;
+      : type = GutterType.medium,
+        size = null;
 
   /// Creates a new [Gutter] widget with a large size.
   const Gutter.large({super.key, this.scaleFactor = 1})
-      : size = null,
-        type = GutterType.large;
+      : type = GutterType.large,
+        size = null;
 
   /// Creates a new [Gutter] widget with an extra large size.
   const Gutter.extraLarge({super.key, this.scaleFactor = 1})
-      : size = null,
-        type = GutterType.extraLarge;
+      : type = GutterType.extraLarge,
+        size = null;
 
   /// Creates a new [Gutter] widget with an expand size.
   const Gutter.expand({super.key, this.scaleFactor = 1})
-      : size = null,
-        type = GutterType.expand;
+      : type = GutterType.expand,
+        size = null;
 
   @override
   Widget build(BuildContext context) {
-    return AxisAware(
+    return AxisAwareOrientation(
       key: key,
       builder: (BuildContext context, Orientation orientation) {
-        final double gapSize =
-            calculateGapSize(context, size, type, scaleFactor);
+        final double gutterSize = context.gutterSize(
+          type: type,
+          size: size,
+          scaleFactor: scaleFactor,
+        );
 
         if (type == GutterType.expand) {
           return Expanded(
             child: SizedBox(
-              width: orientation == Orientation.landscape ? gapSize : null,
-              height: orientation != Orientation.portrait ? null : gapSize,
+              width: orientation == Orientation.landscape ? gutterSize : null,
+              height: orientation != Orientation.portrait ? null : gutterSize,
             ),
           );
         } else {
           return SizedBox(
-            width: orientation == Orientation.landscape ? gapSize : null,
-            height: orientation != Orientation.portrait ? null : gapSize,
+            width: orientation == Orientation.landscape ? gutterSize : null,
+            height: orientation != Orientation.portrait ? null : gutterSize,
           );
         }
       },
@@ -98,16 +104,9 @@ class Gutter extends StatelessWidget {
   }
 }
 
-/// calculates the gap size based on the size, type and scale factor.
-double calculateGapSize(
-    BuildContext context, double? size, GutterType type, double scaleFactor) {
-  return size != null
-      ? (size * scaleFactor)
-      : context.gutter(type: type, scaleFactor: scaleFactor);
-}
-
 /// A gap a quarter the standard gutter size according to Material Design's
 /// breakpoints system
+@Deprecated('Use Gutter.tiny instead')
 class GutterTiny extends Gutter {
   /// Creates a new [GutterTiny] widget.
   const GutterTiny({super.key}) : super.tiny();
@@ -115,6 +114,7 @@ class GutterTiny extends Gutter {
 
 /// A gap half the standard gutter size according to Material Design's
 /// breakpoints system
+@Deprecated('Use Gutter.small instead')
 class GutterSmall extends Gutter {
   /// Creates a new [GutterSmall] widget.
   const GutterSmall({super.key}) : super.small();
@@ -122,176 +122,8 @@ class GutterSmall extends Gutter {
 
 /// A gap twice the standard gutter size according to Material Design's
 /// breakpoints system.
+@Deprecated('Use Gutter.large instead')
 class GutterLarge extends Gutter {
   /// Creates a new [GutterLarge] widget.
   const GutterLarge({super.key}) : super.large();
-}
-
-/// A gap four times the standard gutter size according to Material Design's
-/// breakpoints system.
-class GutterExtraLarge extends Gutter {
-  /// Creates a new [GutterExtraLarge] widget.
-  const GutterExtraLarge({super.key}) : super.extraLarge();
-}
-
-/// A gap of the standard margin size according to Material Design's breakpoints
-/// system
-class Margin extends Gutter {
-  /// Creates a new [Margin] widget.
-  const Margin({super.key, super.size, super.type, super.scaleFactor});
-}
-
-/// A widget that insets its child by the material padding or the given padding.
-class GutterPadding extends Gutter {
-  /// Creates a new [GutterPadding] widget with custom padding.
-  const GutterPadding({
-    super.key,
-    this.child,
-    super.size,
-    super.type = GutterType.medium,
-    super.scaleFactor = 1,
-  })  : left = null,
-        top = null,
-        right = null,
-        bottom = null;
-
-  /// Creates a new [GutterPadding] widget with specific edge insets.
-  const GutterPadding.only({
-    super.key,
-    this.child,
-    this.left = 0,
-    this.top = 0,
-    this.right = 0,
-    this.bottom = 0,
-    super.type = GutterType.medium,
-    super.scaleFactor = 1,
-  });
-
-  /// Creates a new [GutterPadding] widget with symmetric padding.
-  const GutterPadding.symmetric({
-    super.key,
-    this.child,
-    double? vertical = 0,
-    double? horizontal = 0,
-    super.type = GutterType.medium,
-    super.scaleFactor = 1,
-  })  : left = horizontal,
-        top = vertical,
-        right = horizontal,
-        bottom = vertical;
-
-  /// Creates a new [GutterPadding] widget with uniform padding.
-  const GutterPadding.all(
-    double? padding, {
-    super.key,
-    this.child,
-    super.type = GutterType.medium,
-    super.scaleFactor = 1,
-  })  : left = padding,
-        top = padding,
-        right = padding,
-        bottom = padding;
-
-  /// The child widget.
-  final Widget? child;
-
-  /// The offset from the left.
-  final double? left;
-
-  /// The offset from the right.
-  final double? right;
-
-  /// The offset from the top.
-  final double? top;
-
-  /// The offset from the bottom.
-  final double? bottom;
-
-  @override
-  Widget build(BuildContext context) {
-    final double gapSize = calculateGapSize(context, size, type, scaleFactor);
-
-    return Padding(
-      padding: EdgeInsets.only(
-        left: left ?? gapSize,
-        top: top ?? gapSize,
-        right: right ?? gapSize,
-        bottom: bottom ?? gapSize,
-      ),
-      child: child,
-    );
-  }
-}
-
-/// An inherited widget that sets custom [Gutter] and [Gap] behavior.
-class GutterConfigurationData {
-  /// Creates a new [GutterConfigurationData] instance.
-  const GutterConfigurationData({
-    this.widgetToAxis,
-    this.dynamicAxisCheck = true,
-  });
-
-  /// A function to get the axis of a widget not supported by flutter_gutter.
-  ///
-  /// This should return an `Axis` if a widget is recognized and should return
-  /// null if it is not.
-  ///
-  /// Note that this is not necessary if your widget already has an
-  /// `Axis.direction` attribute as Flutter Gutter will find the axis from this
-  /// automatically.
-  ///
-  /// ```dart
-  /// widgetToAxis: (widget) {
-  ///    if (widget is MyCustomHorizontalWidget) {
-  ///     return Axis.horizontal;
-  ///   }
-  ///   return null;
-  /// },
-  /// ```
-  final Axis? Function(Widget widget)? widgetToAxis;
-
-  /// Whether or not gutter widgets should dynamically check for an `axis`
-  /// argument on unrecognized widgets.
-  ///
-  /// Turning this off is useful if you have breakpoints on uncaught exceptions
-  /// on as the dynamic check relies on a try...catch. However, widgets from
-  /// other packages (eg `Boxy`) will not be supported out of the box if you
-  /// turn off the dynamic check.
-  ///
-  /// Defaults to true.
-  final bool dynamicAxisCheck;
-
-  @override
-  bool operator ==(Object other) {
-    return other is GutterConfigurationData &&
-        other.widgetToAxis == widgetToAxis;
-  }
-
-  @override
-  int get hashCode => widgetToAxis.hashCode;
-}
-
-/// An inherited widget that sets custom [Gutter] and [GutterPadding] behavior.
-class GutterConfiguration extends InheritedWidget {
-  /// Creates a new [GutterConfiguration] instance.
-  const GutterConfiguration({
-    super.key,
-    required super.child,
-    required this.data,
-  });
-
-  /// The configuration data.
-  final GutterConfigurationData data;
-
-  /// Gets the [GutterConfigurationData] from the nearest [GutterConfiguration]
-  static GutterConfigurationData? maybeOf(BuildContext context) {
-    return context
-        .dependOnInheritedWidgetOfExactType<GutterConfiguration>()
-        ?.data;
-  }
-
-  @override
-  bool updateShouldNotify(GutterConfiguration oldWidget) {
-    return oldWidget.data == data;
-  }
 }