+++ CHANGELOG.md
@@ -8,8 +8,9 @@
 - Add extension to add Gutter on Lists
 - Migrate to `flutter_adaptive_scaffold` to remove deprecated package
 - Add `AdaptiveGutter` that switches `Gutter` size based on the current breakpoint
-- Add `PaddingGutter` that adds material padding
+- Add `PaddingGutter` that adds material padding to a child widget
 - BREAKING CHANGE: Removal of `Gap`
+- BREAKING CHANGE: Rename `Margin` to `GutterMargin`
 
 ## [1.0.3] - 07/27/2024